+++ storage_drivers/ontap/ontap_san.go
@@ -222,6 +222,7 @@ func (d *SANStorageDriver) Create(
 	exportPolicy := utils.GetV(opts, "exportPolicy", storagePool.InternalAttributes[ExportPolicy])
 	securityStyle := utils.GetV(opts, "securityStyle", storagePool.InternalAttributes[SecurityStyle])
 	encryption := utils.GetV(opts, "encryption", storagePool.InternalAttributes[Encryption])
+	tieringPolicy := utils.GetV(opts, "tieringPolicy", storagePool.InternalAttributes[TieringPolicy])
 
 	if _, _, checkVolumeSizeLimitsError := drivers.CheckVolumeSizeLimits(sizeBytes, d.Config.CommonStorageDriverConfig); checkVolumeSizeLimitsError != nil {
 		return checkVolumeSizeLimitsError
@@ -241,6 +242,10 @@ func (d *SANStorageDriver) Create(
 		storagePool.InternalAttributes[FileSystemType]), name)
 	if err != nil {
 		return err
+	}
+
+	if tieringPolicy == "" {
+		tieringPolicy = d.API.TieringPolicyValue()
 	}
 
 	log.WithFields(log.Fields{
@@ -274,7 +279,7 @@ func (d *SANStorageDriver) Create(
 		// Create the volume
 		volCreateResponse, err := d.API.VolumeCreate(
 			name, aggregate, size, spaceReserve, snapshotPolicy, unixPermissions,
-			exportPolicy, securityStyle, enableEncryption, snapshotReserveInt)
+			exportPolicy, securityStyle, tieringPolicy, enableEncryption, snapshotReserveInt)
 
 		if err = api.GetError(volCreateResponse, err); err != nil {
 			if zerr, ok := err.(api.ZapiError); ok {